requirements.txt
streamlit==1.37.1
langgraph==0.2.33
httpx==0.27.2
tenacity==8.5.0
pydantic==2.9.2
redis==5.0.7  # Placeholder, possible caching backend
python-dotenv==1.0.1
tenacity==8.5.0
tldextract==5.1.2
openai==1.51.2
orjson==3.10.7 # structured logging